% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prostar.R
\name{Prostar}
\alias{Prostar}
\title{Prostar}
\usage{
Prostar()
}
\value{
A new window in the default internet browser
}
\description{
Prostar
}
\examples{
if (interactive()) {
    Prostar()
}

}
\author{
Samuel Wieczorek
}
